requirements.txt
aiohttp==3.11.18
beautifulsoup4==4.13.4
duckduckgo_search==8.0.2
exa_py==1.14.10
fastapi==0.115.13
google-generativeai==0.8.3
httpx==0.28.1
langchain==0.3.26
langchain_anthropic==0.3.15
langchain_community==0.3.26
langchain_core==0.3.66
langgraph==0.4.8
langsmith==0.4.1
linkup==0.1.3
linkup_sdk==0.2.5
markdownify==1.1.0
openai==1.90.0
pydantic==2.11.7
python-dotenv==1.1.0
Requests==2.32.4
supabase==2.15.3
tavily_python==0.7.2
uvicorn==0.34.3
